// src/types/storyblok.d.ts
// This file was generated by the storyblok CLI.
// DO NOT MODIFY THIS FILE BY HAND.
import type { ISbStoryData } from "storyblok";
export interface FeatureStoryblok {
  headline: string;
  content: string;
  variant: "" | "primary" | "secondary" | "tertiary";
  component: "feature";
  _uid: string;
  [k: string]: any;
}

export interface GridStoryblok {
  headline?: string;
  items: (FeatureStoryblok | TestimonialStoryblok)[];
  component: "grid";
  _uid: string;
  [k: string]: any;
}

export interface AssetStoryblok {
  alt: string | null;
  copyright?: string | null;
  fieldtype: "asset";
  id: number;
  filename: string | null;
  name: string;
  title: string | null;
  focus: string | null;
  meta_data?: {
    [k: string]: any;
  };
  source?: string | null;
  is_external_url?: boolean;
  is_private?: boolean;
  src?: string;
  updated_at?: string;
  width?: number | null;
  height?: number | null;
  aspect_ratio?: number | null;
  public_id?: string | null;
  content_type?: string;
  [k: string]: any;
}

export interface HeroStoryblok {
  headline: string;
  content: string;
  image: AssetStoryblok;
  component: "hero";
  _uid: string;
  [k: string]: any;
}

export interface PageStoryblok {
  blocks?: (FeatureStoryblok | GridStoryblok | HeroStoryblok | RecommendedRecipesStoryblok)[];
  component: "page";
  _uid: string;
  [k: string]: any;
}

export interface RichtextStoryblok {
  type: string;
  content?: RichtextStoryblok[];
  marks?: RichtextStoryblok[];
  attrs?: any;
  text?: string;
  [k: string]: any;
}

export interface RecipeStoryblok {
  body: RichtextStoryblok;
  name: string;
  introduction: string;
  preparation_time: string;
  main_img: AssetStoryblok;
  category: "" | "Meat" | "Dessert" | "Salad" | "Soup" | "Pastry" | "Pasta" | "Vegeterian";
  servings: string;
  rating: string;
  component: "recipe";
  _uid: string;
  [k: string]: any;
}

export interface RecommendedRecipesStoryblok {
  headline: string;
  recipes: (ISbStoryData<RecipeStoryblok> | string)[];
  component: "recommended_recipes";
  _uid: string;
  [k: string]: any;
}

export interface TestimonialStoryblok {
  comment: string;
  name: string;
  image: AssetStoryblok;
  component: "testimonial";
  _uid: string;
  [k: string]: any;
}